renames testUtils to devUtils
