+++ ironfish/src/index.ts
@@ -28,4 +28,4 @@ export * from './package'
 export * from './platform'
 export * from './primitives'
 export { getFeeRate } from './memPool'
-export * as testUtils from './testUtils'
+export * as devUtils from './devUtils'
