+++ ironfish/src/multisig.test.slow.ts
@@ -5,7 +5,7 @@
 import { Asset, multisig, Note as NativeNote, verifyTransactions } from '@ironfish/rust-nodejs'
 import { Note, RawTransaction } from './primitives'
 import { Transaction, TransactionVersion } from './primitives/transaction'
-import { makeFakeWitness } from './testUtils'
+import { makeFakeWitness } from './devUtils'
 
 describe('multisig', () => {
   describe('dkg', () => {
